Added BankService

+++ BankAPI/Controllers/SystemController.cs
@@ -16,14 +16,14 @@ namespace BankAPI.Controllers
         [HttpGet]
         public ActionResult GetSystemVariable()
         {
-            var system = this.db.SystemVariables.Find(1);
+            var system = this.db.SystemVariables.Find((byte)1);
             return Ok(new SystemVariableDTO() { CurrentDate = system.CurrentDate });
         }
 
         [HttpPost]
         public ActionResult UpdateSystemVariable([FromBody] UpdateSystemRequest request)
-        {
-            var system = this.db.SystemVariables.Find(1);
+        { 
+            var system = this.db.SystemVariables.Find((byte)1);
 
             system.CurrentDate = request.CurrentDate;
